+++ src/lightning/fabric/_graveyard/tpu.py
@@ -15,6 +15,7 @@ import sys
 from typing import Any
 
 import lightning.fabric as fabric
+from lightning.fabric.accelerators import XLAAccelerator
 from lightning.fabric.strategies import _StrategyRegistry
 from lightning.fabric.strategies.single_xla import SingleDeviceXLAStrategy
 from lightning.fabric.utilities.rank_zero import rank_zero_deprecation
@@ -23,12 +24,13 @@ from lightning.fabric.utilities.rank_zero import rank_zero_deprecation
 def _patch_sys_modules() -> None:
     self = sys.modules[__name__]
     sys.modules["lightning.fabric.strategies.single_tpu"] = self
+    sys.modules["lightning.fabric.accelerators.tpu"] = self
 
 
 class SingleTPUStrategy(SingleDeviceXLAStrategy):
     """Legacy class.
 
-    Use `SingleDeviceXLAStrategy` instead.
+    Use :class:`~lightning.fabric.strategies.single_xla.SingleDeviceXLAStrategy` instead.
     """
 
     def __init__(self, *args: Any, **kwargs: Any):
@@ -40,8 +42,22 @@ class SingleTPUStrategy(SingleDeviceXLAStrategy):
         strategy_registry.register("single_tpu", cls, description="Legacy class. Use `single_xla` instead.")
 
 
+class TPUAccelerator(XLAAccelerator):
+    """Legacy class.
+
+    Use :class:`~lightning.fabric.accelerators.xla.XLAAccelerator` instead.
+    """
+
+    def __init__(self, *args: Any, **kwargs: Any):
+        rank_zero_deprecation(
+            "The `TPUAccelerator` class is deprecated. Use `lightning.fabric.accelerators.XLAAccelerator` instead."
+        )
+        super().__init__(*args, **kwargs)
+
+
 def _patch_classes() -> None:
     setattr(fabric.strategies, "SingleTPUStrategy", SingleTPUStrategy)
+    setattr(fabric.accelerators, "TPUAccelerator", TPUAccelerator)
 
 
 _patch_sys_modules()