+++ src/lightning/fabric/accelerators/xla.py
@@ -13,17 +13,18 @@
 # limitations under the License.
 import functools
 from multiprocessing import Process, Queue
-from typing import Any, Callable, Dict, List, Union
+from typing import Any, Callable, List, Union
 
 import torch
 from lightning_utilities.core.imports import RequirementCache
 
+from lightning.fabric.accelerators import _AcceleratorRegistry
 from lightning.fabric.accelerators.accelerator import Accelerator
 from lightning.fabric.utilities.device_parser import _check_data_type
 
 
-class TPUAccelerator(Accelerator):
-    """Accelerator for TPU devices.
+class XLAAccelerator(Accelerator):
+    """Accelerator for XLA devices, normally TPUs.
 
     .. warning::  Use of this accelerator beyond import and instantiation is experimental.
     """
@@ -99,12 +100,8 @@ class TPUAccelerator(Accelerator):
         return queue.get_nowait()
 
     @classmethod
-    def register_accelerators(cls, accelerator_registry: Dict) -> None:
-        accelerator_registry.register(
-            "tpu",
-            cls,
-            description=cls.__class__.__name__,
-        )
+    def register_accelerators(cls, accelerator_registry: _AcceleratorRegistry) -> None:
+        accelerator_registry.register("tpu", cls, description=cls.__class__.__name__)
 
 
 # define TPU availability timeout in seconds
@@ -160,7 +157,7 @@ def _parse_tpu_devices(devices: Union[int, str, List[int]]) -> Union[int, List[i
 
 
 def _check_tpu_devices_valid(devices: object) -> None:
-    device_count = TPUAccelerator.auto_device_count()
+    device_count = XLAAccelerator.auto_device_count()
     if (
         # support number of devices
         isinstance(devices, int)