+++ tests/tests_pytorch/graveyard/test_tpu.py
@@ -17,3 +17,17 @@ def test_graveyard_single_tpu(import_path, name):
     device = torch.device("cpu")
     with pytest.deprecated_call(match="is deprecated"), pytest.raises(ModuleNotFoundError, match="torch_xla"):
         cls(device)
+
+
+@pytest.mark.parametrize(
+    ("import_path", "name"),
+    [
+        ("lightning.pytorch.accelerators", "TPUAccelerator"),
+        ("lightning.pytorch.accelerators.tpu", "TPUAccelerator"),
+    ],
+)
+def test_graveyard_tpuaccelerator(import_path, name):
+    module = import_module(import_path)
+    cls = getattr(module, name)
+    with pytest.deprecated_call(match="is deprecated"), pytest.raises(ModuleNotFoundError, match="torch_xla"):
+        cls()